Add auth to retrain and get price endpoints
