--- bin/receipts.metadata.json
@@ -1,1 +1,0 @@
-{"options":{},"indexes":[{"v":2,"key":{"_id":1},"name":"_id_","ns":"porton.receipts"}]}
