--- bin/receipts_dict.json
@@ -1,1 +1,0 @@
-{"type": {"004": 3, "001": 0, "006": 5, "002": 1, "005": 2, "007": 4}, "specialty": {"Neurology": 1, "Cardiology": 0, "Dermatology": 2, "Pediatrics": 11, "Family Medicine/General Practitioner": 9, "Plastic Surgery": 10, "Allergy and Immunology": 6, "Aerospace Medicine": 7, "Clinical Counselling": 8, "Pulmonology": 4, "Anesthesiology": 5, "Clinical Genetics and Genomics": 3}, "eventType": {"report": 2, "voice": 1, "video": 0}}
