--- receipts_schema.json
@@ -1,6 +1,0 @@
-{
-	
-	"specialty": "class",
-	"eventType": "class",
-	"type": "class"
-}
