+++ server/bin/receipts.metadata.json
@@ -1,0 +1,1 @@
+{"options":{},"indexes":[{"v":2,"key":{"_id":1},"name":"_id_","ns":"porton.receipts"}]}
