+++ server/bin/receipts_dict.json
@@ -1,0 +1,1 @@
+{"specialty": {"Cardiology": 0, "Neurology": 1, "Dermatology": 2, "Clinical Genetics and Genomics": 3, "Pulmonology": 4, "Anesthesiology": 5, "Allergy and Immunology": 6, "Aerospace Medicine": 7, "Clinical Counselling": 8, "Family Medicine/General Practitioner": 9, "Plastic Surgery": 10, "Pediatrics": 11}, "eventType": {"video": 0, "voice": 1, "report": 2}, "type": {"001": 0, "002": 1, "005": 2, "004": 3, "007": 4, "006": 5}}
