+++ server/receipts_schema.json
@@ -1,0 +1,6 @@
+{
+	
+	"specialty": "class",
+	"eventType": "class",
+	"type": "class"
+}
